+++ openmls/src/group/core_group/create_commit.rs
@@ -50,7 +50,7 @@ impl CoreGroup {
                 // group.
                 let leaf_index =
                     self.free_leaf_index(params.inline_proposals().iter().map(Some))?;
-                (Sender::build_new_member(), leaf_index)
+                (Sender::NewMemberCommit, leaf_index)
             }
             CommitType::Member => (
                 Sender::build_member(
@@ -184,7 +184,7 @@ impl CoreGroup {
             };
 
         let sender = match params.commit_type() {
-            CommitType::External => Sender::build_new_member(),
+            CommitType::External => Sender::NewMemberCommit,
             CommitType::Member => Sender::build_member(
                 self.key_package_ref()
                     .ok_or_else(|| LibraryError::custom(" missing key package"))?,