+++ openmls/src/messages/mod.rs
@@ -27,6 +27,7 @@ use tls_codec::{Serialize as TlsSerializeTrait, *};
 
 // Public
 pub mod codec;
+pub mod external_proposals;
 pub mod proposals;
 pub mod public_group_state;